one_l/agent_api/functions/agent/document_review/requirements.txt
boto3>=1.34.0
python-docx>=1.1.0
anthropic>=0.34.0